+++ packages/flutter_riverpod/lib/flutter_riverpod.dart
@@ -1,5 +1,5 @@
 // ignore: invalid_export_of_internal_element, Already tackled by riverpod/riverpod.dart. If we export internals, that's on purpose.
 export 'package:riverpod/riverpod.dart';
 
-export 'src/change_notifier_provider.dart';
 export 'src/core.dart' hide ProviderScopeState;
+export 'src/providers/legacy/change_notifier_provider.dart';
