+++ packages/flutter_riverpod/lib/src/internals.dart
@@ -1,3 +1,4 @@
 export 'package:riverpod/src/internals.dart';
-export 'change_notifier_provider.dart';
+
 export 'core.dart';
+export 'providers/legacy/change_notifier_provider.dart';
